household-service/pyproject.toml
[project]
name = "household-service"
version = "0.1.0"
description = ""
authors = [
    {name = "Kai Deines",email = "deines.kl@outlook.de"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi (>=0.135.3,<0.136.0)",
    "uvicorn (>=0.43.0,<0.44.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "sqlalchemy (>=2.0.49,<3.0.0)"
]

[tool.poetry]
packages = [{include = "household_service", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
